Lista Prove funzionante

+++ AreaDocente/Models/Entities/MVCTest_DA.cs
@@ -8,7 +8,7 @@ namespace AreaDocente.Models.Entities
     {
         [Key]
         public Guid K_Test_DA { get; set; }
-        public int? Numero_Domanda { get; set; }
+        public byte? Numero_Domanda { get; set; }
         public string? Domanda { get; set; }
         public string? Risposta { get; set; }
         public Guid? K_Prova { get; set; }
